Error via exception + sector action
Error are now returned via exception (as we should always do).
Sector action are now implemented.

+++ sources/Doom/Camera.hpp
@@ -42,7 +42,7 @@ namespace DOOM
     bool	renderNode(const DOOM::Doom & doom, int16_t index);																			// Render level BSP tree recursively from its root node
     bool	renderSubsector(const DOOM::Doom & doom, int16_t index);																		// Iterate through seg of subsector
     bool	renderSeg(const DOOM::Doom & doom, int16_t index);																			// Projection of segment on screen
-    void	renderTexture(const DOOM::Doom & doom, const DOOM::Doom::Resources::Texture & texture, int column, float top, float bottom, float height, int offset_x, int offset_y, int16_t light, int16_t seg);	// Draw a column from a texture
+    void	renderTexture(const DOOM::Doom & doom, const DOOM::Doom::Resources::Texture & texture, int column, float top, float bottom, float height, int offset_x, float offset_y, int16_t light, int16_t seg);	// Draw a column from a texture
     void	renderFlat(const DOOM::Doom & doom, const DOOM::AbstractFlat & flat, int column, int start, int end, float altitude, int16_t light, int16_t seg);							// Draw a column from a flat
     void	renderSky(const DOOM::Doom & doom, int column, int start, int end, float altitude, int16_t seg);													// Draw a column from a sky texture
     void	renderThings(const DOOM::Doom & doom);																					// Draw things of current level
